chat

+++ server/common/actor_type.go
@@ -39,7 +39,7 @@ func WorldName(id int32) ActorId {
 func GatewayName(id int32) ActorId {
 	return fmt.Sprintf("%v_%v_Actor", GateWay_Actor, id)
 }
-func ChatName(id int32) ActorId {
+func ChatName(id uint16) ActorId {
 	return fmt.Sprintf("%v_%v_Actor", Chat_Actor, id)
 }
